+++ fuzzers/qsym/fuzzer.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 """Integration code for QSYM fuzzer."""
 
-import contextlib
 import shutil
 import subprocess
 import os
@@ -43,7 +42,7 @@ def build():
 
     src = os.getenv('SRC')
     work = os.getenv('WORK')
-    with restore_directory(src), restore_directory(work):
+    with utils.restore_directory(src), utils.restore_directory(work):
         # Restore SRC to its initial state so we can build again without any
         # trouble. For some OSS-Fuzz projects, build_benchmark cannot be run
         # twice in the same directory without this.
@@ -112,37 +111,3 @@ def fuzz(input_corpus, output_corpus, target_binary):
         './qsym/bin/run_qsym_afl.py', '-a', 'afl-slave', '-o', output_corpus,
         '-n', 'qsym', '--', uninstrumented_target_binary
     ])
-
-
-@contextlib.contextmanager
-def restore_directory(directory):
-    """Helper contextmanager that when created saves a backup of |directory| and
-    when closed/exited replaces |directory| with the backup.
-
-    Example usage:
-
-    directory = 'my-directory'
-    with restore_directory(directory):
-       shutil.rmtree(directory)
-    # At this point directory is in the same state where it was before we
-    # deleted it.
-    """
-    # TODO(metzman): Figure out if this is worth it, so far it only allows QSYM
-    # to compile bloaty.
-    if not directory:
-        # Don't do anything if directory is None.
-        yield
-        return
-    # Save cwd so that if it gets deleted we can just switch into the restored
-    # version without code that runs after us running into issues.
-    initial_cwd = os.getcwd()
-    with tempfile.TemporaryDirectory() as temp_dir:
-        backup = os.path.join(temp_dir, os.path.basename(directory))
-        shutil.copytree(directory, backup)
-        yield
-        shutil.rmtree(directory)
-        shutil.move(backup, directory)
-        try:
-            os.getcwd()
-        except FileNotFoundError:
-            os.chdir(initial_cwd)
