+++ fuzzers/utils.py
@@ -13,11 +13,46 @@
 # limitations under the License.
 """Utility functions for running fuzzers."""
 
+import contextlib
 import os
 import shutil
 import subprocess
 
 OSS_FUZZ_LIB_FUZZING_ENGINE_PATH = '/usr/lib/libFuzzingEngine.a'
+
+
+@contextlib.contextmanager
+def restore_directory(directory):
+    """Helper contextmanager that when created saves a backup of |directory| and
+    when closed/exited replaces |directory| with the backup.
+
+    Example usage:
+
+    directory = 'my-directory'
+    with restore_directory(directory):
+       shutil.rmtree(directory)
+    # At this point directory is in the same state where it was before we
+    # deleted it.
+    """
+    # TODO(metzman): Figure out if this is worth it, so far it only allows QSYM
+    # to compile bloaty.
+    if not directory:
+        # Don't do anything if directory is None.
+        yield
+        return
+    # Save cwd so that if it gets deleted we can just switch into the restored
+    # version without code that runs after us running into issues.
+    initial_cwd = os.getcwd()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        backup = os.path.join(temp_dir, os.path.basename(directory))
+        shutil.copytree(directory, backup)
+        yield
+        shutil.rmtree(directory)
+        shutil.move(backup, directory)
+        try:
+            os.getcwd()
+        except FileNotFoundError:
+            os.chdir(initial_cwd)
 
 
 def build_benchmark(env=None):
